Change money/amount properties from int to decimal
Every monetary field across the request/response models (Amount,
Fee, AppFee, MerchantFee, AvailableBalance, AmountRefunded, etc.) was
typed as int, silently truncating fractional currency values returned
by or sent to the API. Change them to decimal throughout, in both the
external DTOs and the public domain models, and fix the one Fee field
that was float for the same reason.

Fields that only look money-shaped but aren't (pagination counts,
retry-strategy config values, charge counters, IDs) were left as int.

This surfaced a real bug: several *Validations.cs files only had an
IsInvalid(double) overload for numeric checks, so once a validated
field's static type became decimal, `IsInvalid(field)` silently
rebound to IsInvalid(object) (decimal boxes to a non-null object) and
stopped flagging non-positive amounts as invalid. Added matching
IsInvalid(decimal) overloads (and IsInvalid(int) where needed to
resolve now-ambiguous calls from plain int parameters) everywhere the
double overload existed, restoring the original validation behavior.

+++ FlutterWave.Core/Models/Services/Foundations/ExternalFlutterWave/ExternalBillPayments/ExternalBillStatusPaymentResponse.cs
@@ -19,10 +19,10 @@ namespace FlutterWave.Core.Models.Services.Foundations.ExternalFlutterWave.Exter
             public string TxRef { get; set; }
 
             [JsonProperty("amount")]
-            public int Amount { get; set; }
+            public decimal Amount { get; set; }
 
             [JsonProperty("fee")]
-            public int Fee { get; set; }
+            public decimal Fee { get; set; }
 
             [JsonProperty("currency")]
             public object Currency { get; set; }
