+++ FlutterWave.Core/Models/Services/Foundations/FlutterWave/Charge/ApplePayResponse.cs
@@ -48,7 +48,7 @@ namespace FlutterWave.Core.Models.Services.Foundations.FlutterWave.Charge
             public string DeviceFingerprint { get; set; }
 
             [JsonProperty("amount")]
-            public int Amount { get; set; }
+            public decimal Amount { get; set; }
 
             [JsonProperty("charged_amount")]
             public double ChargedAmount { get; set; }
@@ -57,7 +57,7 @@ namespace FlutterWave.Core.Models.Services.Foundations.FlutterWave.Charge
             public double AppFee { get; set; }
 
             [JsonProperty("merchant_fee")]
-            public int MerchantFee { get; set; }
+            public decimal MerchantFee { get; set; }
 
             [JsonProperty("processor_response")]
             public string ProcessorResponse { get; set; }
