+++ FlutterWave.Core/Models/Services/Foundations/FlutterWave/Misc/AllBalanceCurrenciesResponse.cs
@@ -11,7 +11,7 @@ namespace FlutterWave.Core.Models.Services.Foundations.FlutterWave.Miscellaneous
         public class Datum
         {
             public string Currency { get; set; }
-            public int AvailableBalance { get; set; }
+            public decimal AvailableBalance { get; set; }
             public double LedgerBalance { get; set; }
         }
